+++ src/components/form/password-field/PasswordField.tsx
@@ -3,7 +3,7 @@ import TextField from '../text-field'
 import { PasswordFieldProps } from './types'
 import { TextFieldProps } from '../text-field/types'
 
-const PasswordField: React.FC<TextFieldProps & PasswordFieldProps> = ({ label }) => {
+const PasswordField: React.FC<TextFieldProps & PasswordFieldProps> = ({ label, ...otherProps }) => {
     const [opened, setOpened] = React.useState(false)
     const handleAction = () => setOpened(!opened)
 
@@ -14,6 +14,7 @@ const PasswordField: React.FC<TextFieldProps & PasswordFieldProps> = ({ label })
             actionIcon={opened ? 'eye-slash' : 'eye'}
             onActionTriggered={handleAction}
             isPassword={!opened}
+            {...otherProps}
         />
     )
 }