+++ src/components/form/text-field/types.ts
@@ -1,4 +1,5 @@
 import { IconNameType } from '../../icon/types'
+
 export interface TextFieldProps {
     label?: string
     secondary?: boolean
@@ -12,4 +13,5 @@ export interface TextFieldProps {
     actionIcon?: IconNameType
     onActionTriggered?: () => void
     isPassword?: boolean
+    autoCapitalize?: 'characters' | 'words' | 'sentences' | 'none'
 }