+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [added]: faGlassWater icon
 - [added]: faClock icon
 - [added]: faPlay icon
+- [added]: autoCapitalize prop to text input